Add filetype blacklist
Exist languages with significant trailing whitespace, e.g. Markdown.
Adds the flag `b:strip_trailing_whitespace_enabled` for selectively
disabling the plugin on a per-buffer basis. The tracking of modified
lines still always has to happen though, to support re-enabling the
plugin.

In the autocmd group `strip_trailing_whitespace_filetype` an autocommand
is added for the `FileType` event that blacklists diff and Markdown
files and activates the plugin for other filetypes. To get rid of the
default behaviour:

    :autocmd! strip_trailing_whitespace_filetype

Alternatively an explicit global variable could have been added as a
more easily customizable blacklist. However, since a buffer toggle is
warranted also for one-off situations, this was not preferred.